Separate pre only providers from the staking provider list
